fix tick / deltatime behaviour in userscripts
also theoretically added deserialization of scene files but it's broken + BinaryFormatter approach has been abandoned and will be switched out for either XML or custom formatting

+++ Tyme Engine/Tyme Engine/EngineSource/Core/UserScript.cs
@@ -9,6 +9,6 @@ namespace Tyme_Engine.Core
     class UserScript : Component
     {
         public virtual void Start(){}
-        public virtual void Tick(double delta){}
+        public virtual void Tick(float delta){}
     }
 }
